Refactor BuildXPrebidHeader Method
